requirements.txt
kiteconnect>=5.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
